internal/counter: remove unused build tag
The bootstrap build tag was needed when this package was
in the go project - so instrumentation doesn't interfere
building bootstrap binaries.

This package is outside of the go main project, so this tag
is not necessary. When the x/telemetry/counter and its
dependencies are moved to the go project, we will add the
tags in appropriate places.

Change-Id: Ifdb615b201d2ddeef7e0fbe71db123d16a8590d0
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/505719
Reviewed-by: Jamal Carvalho <jamal@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Hyang-Ah Hana Kim <hyangah@gmail.com>