internal/counter: compress stack counter names
Stack counter names are multiline, where each line is
<import-path>.<function-name>:<line-number>. The import paths
tend to be long and repeated, making the names stored in count files
significantly longer than necessary. This change replaces
a repeated import path in the sequence of lines with a single
character " (for 'Ditto'). This is a detail of the internals
of the count files. Parse() reconstructs the uncompressed names.

For: golang/go#61234

Change-Id: I1f1a4160cf7b16a1e2937b56d8c929736bb6ea5c
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/509435
Run-TryBot: Peter Weinberger <pjw@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Jamal Carvalho <jamal@golang.org>