internal/counter: fix counter.Add behavior before having counter pointer
Before the counter data pointer is set, counter.Add stores the value
in the extra bits, by reading the current extra and add the value
and again assigning the sume to the state. This is done with bit
operation and we forgot to clear the previous extra bits before or-ing
the new value.

We update counterStateBits using compare-and-swap.
If CAS fails, we keep retrying in a loop. Unfortunately, we forgot to
load the current state before the next iteration. Fix this.

Change-Id: I9da738f56f70bd5e88294537d5f0caf011c09268
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/513818
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Hyang-Ah Hana Kim <hyangah@gmail.com>
Reviewed-by: Jamal Carvalho <jamal@golang.org>
Reviewed-by: Russ Cox <rsc@golang.org>