internal/upload: generate reports and remove count files
To limit the chance for a race when two processes are simultaneously
uploading, the code checks to see if there is already a report just
before writing one. This decreases the window to just the time
between a Stat and a Write.

counter.Open return a function to be used to cleanup in tests
on Windows. (Windows will not delete an open file.)

Note that the random probability X is not being used to decide
what to upload, as was in the original specification, when telemetry
was opt-out. This could be changed.

Change-Id: I7693b6213a57986913edea23486dc2ce179ffbd8
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/515696
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Peter Weinberger <pjw@google.com>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>