internal/counter/stackcounter: let ReadStack return matching counters
countertest.ReadStackCounter takes a counter.StackCounter instance
and it returns only the entries that are found in the in-memory
map of the instance.

This is tricky when the stack counter instance is defined in some
other package and access to it is not allowed.

Change-Id: I6b5a847f5361d438ddb883ee0d99c220b5729d30
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/515996
Run-TryBot: Hyang-Ah Hana Kim <hyangah@gmail.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Jamal Carvalho <jamal@golang.org>