telemetry/upload: proposed public interface
This is a proposal for the public upload.Run API. Comments
and suggestions are more than welcome.

Also, consider if it would be a good idea to add a
log.Logger to the Configuration, and if so, what
the default should be when it isn't given. (E.g., what
should the upload package log when used in gopls?)

Change-Id: I99e25c434e12362cfa964352347bd9dc85db914a
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/518375
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Peter Weinberger <pjw@google.com>
Reviewed-by: Jamal Carvalho <jamal@golang.org>