internal/counter: introduce and use weekdays file
and make all dates refer to UTC.

This CL introduces the weekends file in telemetry.LocalDir. The file
contains an ascii time.Weekday (a string in the range "0"..."6" inclusive)
which determines the expiration day of the week for all counter files.
If it does not exist the file is created and filled with a random day
of the week. (There is nothing preventing knowledgeable users from
changing it.)

Before there are any reports (i.e. for a new user) the expiration day
for counter files will be the day the week ends the following week.
Otherwise the expiration day will be the next (1 to 7 days from now)
time the week ends.

Also fix countertest.go, to create temporary dirs.

Change-Id: Ic1a7f286dc4043466f8a71fb9db1568a665201eb
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/520455
Reviewed-by: Jamal Carvalho <jamal@golang.org>
Run-TryBot: Peter Weinberger <pjw@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Robert Findley <rfindley@google.com>