cmd/gotelemetry: local viewer ui cleanup
Charts are moved to the top of the page.

Removed the facet selector as it offers little benefit
and is not intuitive to use.

Added domain to the chartdata output to align the week
intervals for each of the graphs.

Fixed UI for unregistered stack counters.

Change-Id: Id1783f13c9bf175fc744b47d79c9a7cd41cd62e0
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/520675
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Jamal Carvalho <jamal@golang.org>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>