godev/cmd/worker: add chart endpoint
Similar to the merge endpoint, we'll call this with Cloud Scheduler
at some interval.

Currently this job will only produce partition graph data for a single
week and relies on the upload config to determine the shape of the
output. To generate chart data for a merged report make a request to
/charts?date=<date>. A future CL will remove the query param requirement
and instead run the job for a fixed interval of days, according to the time window we allow for report uploads.

Only partition charts are generated at this time. The chart endpoint
must be updated to get the chart type from the graph config.

Change-Id: Ida14498c20bc8276e762f191ade94de9a78e2740
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/523736
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Jamal Carvalho <jamal@golang.org>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>