internal/regtest: remove the need of Main
Requiring to register all functions in TestMain split up the definition
of the program and the call to RunProg too far.

Instead, we provide NewProgram that allows to colocate the definition
and usage of the program within the test function.

func TestFoo(t *testing.T) {
  telemetryDir := t.TempDir()
  prog := regtest.NewProgram(t, "foo", func() int { ... })
  out, err := regtest.RunProg(t, telemetryDir, prog)
  ...
}

Change-Id: I8dc427f4bb5ec3c5853959ed2e70e7681148d539
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/551935
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>