internal/counter: clarify relative line numbers
This CL changes the notation used in stack counters to make
the line numbering scheme more obvious to the reader. We use
"f:+3" to mean "3 lines after func f",
"f:-3" to mean "3 lines before func f", and
"f:=3" to mean "line 3 in the file containing f".

Previously, these were "f:3", "f:?abs" where abs
is the file-relative line number, and "f:??3", all
of which were more confusing to me.

This is a schema change, so we'll be forced to dedup the
new reports, but better to get this in now before the
real flow of crash reports begins.

Also, eliminate unnecessary locals, and remove unnecessary
zeroing of lastImport in f:=%d case.

Change-Id: Ifae4dbd6beee73703529952fa6e7ab1e54c72cd1
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/561495
Reviewed-by: Peter Weinberger <pjw@google.com>
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>