start: add Start which may start crashmonitor
This change starts the implementation of the proposed telemetry.Start
function as in #65500 which is supposed to start the crashmonitor and
uploader depending on the configuration. For simplicity this change
just adds the option to start the crashmonitor. The functionality in
the crashmonitor package is mostly moved to internal/crashmonitor
(except for Supported) in Parent and Child functions that are meant to
be invoked in the parent process and the telemetry child to set up
crash monitoring. We'll do something similar for uploading.

I couldn't find a clean way to put Start in the telemetry package, so
I put it in a start package instead. Because telemetry is at the
bottom of the import graph it would be hard to expose the Start
function there without having a lot of messiness.

For #65500

Change-Id: Ied44a9140b32488120849fb4f2a124dbe968acb9
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/563955
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Alan Donovan <adonovan@google.com>