telemetry: add support for uploading
No support for rate limiting yet but want to keep things small here.
This change adds a dependency on x/sync for errgroup.

I've also had to run go mod tidy in the godev module because it has a
replacement instead of depending on a version of this module.

For golang/go#65500

Change-Id: Icbcd8d3d58fb443c14cf21e339baa3edb76fbca1
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/564615
Reviewed-by: Robert Findley <rfindley@google.com>
Reviewed-by: Bryan Mills <bcmills@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>