telemetry: log to a file from telemetry child instead of stderr
Instead of logging to config.Logger or stderr from the telemetry
child, log to a log file that we open in the telemetry local dir. (Is
that the right place to put it?) This will allow us to close stderr
and not open stdout from the child so it can be better behaved.

Change-Id: I3397c9775a58d3902b4d3636e67976e045ffd6a2
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/566056
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>