telemetry: in Start, run upload.Run at most once a day
This change will create a token file to acquire a token to limit
uploading to at most once a day: before upload is run, we'll check to
see if an upload token file already exists. If it doesn't, or it's
been a day since the existing file has been created, it's been a day
since the last upload.Run was attempted so we can run it again. If it
doesn't, we'll skip running upload.Run. We don't delete the token
files when we're done with the upload, but instead only delete them
when we want to acquire a new lock so that the existence of the file
and its modification time is a signal of the last time uploading was
run.

The token acquiring code is based on acquireLockFile in
x/tools/gopls/internal/server/prompt.go.

For golang/go#65500

Change-Id: Ie31c1f351e2d31a016fd2ad79b29784e6631c564
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/567075
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>