telemetry: add option to supply TelemetryDir
This will allow users of telemetry.Start to specify an alternate
location for the telemetry dir to use when a binary is running as part
of a test. It is meant an alternative for countertest.Open for
programs using telemetry.Start.

For #65500

Change-Id: Ifeba9c6fa53fba99dc259d8a031fe287c7fb58b9
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/569315
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>