internal/upload: close over state when uploading, for easier testing
As described in golang/go#66003 (and illustrated in a recent bug in go
command integration), it is important that downstream tools are able to
write end-to-end tests for uploading. Yet the upload process accessed a
fair amount of shared global state, presenting a challenge for
testability.

Fix this by closing over all state in the Uploader, and expose
configuration hooks in a new upload.RunConfig type. Specifically:

- Rename upload.Control to upload.RunConfig (since it is config for
  Run), and move it into the internal/upload package (with an alias in
  upload), so that it can be used for internal/telemetry tests
- Add fields necessary for testing to RunConfig: TelemetryDir,
  UploadURL, LogWriter, UploadConfig, and StartTime.
- Unexport all fields from Uploader, since hooks are all exposed via
  RunConfig.
- Remove the global upload.logger, since logging is now scoped to
  individual log operations.
- Move initialization of the Uploader into NewUploader, rather than
  delaying the resolution of the UploadConfig or having a separate
  LogIfDebug call. This way, the constructed Uploader is ready to use
  and immutable.
- Let upload.Run (both variants) return an error if the upload failed
  completely.
- Update configstore.Download to return a pointer, for convenience.
- Add the Debug dir to the internal/telemetry.Dir mapper.

Together, these changes expose the seams necessary to test the Upload
(possibly concurrently) from external integrations such as the Go
command. In the future, we may want to reduce the surface area of this
configuration, for example by setting GOPROXY rather than passing in an
upload configuration.

Also:
 - Fix a bug where uploads proceeded even if the config download failed.
   Now, the config is downloaded during NewUploader, which can fail with
   an error.
 - Change the debug directory from "<telemetryDir>/local/debug" to
   "<telemetryDir>/debug", since the nesting felt unnecessary (I don't
   feel strongly about this)

For golang/go#66003

Change-Id: I4125bfe682d11122c5bb2a605717a366108a862d
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/575060
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>