internal/chartconfig: support multiline counter fields
Counter fields are allowed to span multiple lines within the curly
braces used for buckets. If we see an open curly brace on a counter
line start accumulating a multiline counter field, by joining the
lines up to an including the next line with a curly brace. We do error
checking to make sure that curlies are not nested or mismatched.

Change-Id: I5554912dee773c08d46987462068b833aa28338c
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/579915
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>