telemetry: in Start, set X_TELEMETRY_CHILD=2 before running the child
As described in the code, failing to modify X_TELEMETRY_CHILD can lead to
go subcommands acting as the telemetry child themselves, rather than as
the go command. On the other hand, we don't want the delegated go
commands to *also* fork themselves. Therefore, we establish a convention
that X_TELEMETRY_CHILD=1 implies that the process is a telemetry child
process, and X_TELEMETRY_CHILD=2 implies that the process is a child of
a telemetry child process.

For golang/go#67211

Change-Id: Ia6924cc1671f9500f21cc9102cc7d5da33251a83
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/583575
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>