internal/counter: remove fall-back logic in Parse for missing TimeBegin
Remove fallback logic in counter.Parse to handle a missing "TimeBegin"
metadata field. All counters should have this field populated (it is
written by file.init). To ensure we catch this form of corruption,
return an error from counterDateSpan when the TimeBegin or TimeEnd
fields are missing (this was a pre-existing TODO).

The logic dates to the original instance of the counter package, so it
is likely just obsolete.

For golang/go#65970

Change-Id: I6a75a42f3092c3471fe95cda1958b57b954e9140
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/584402
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>
Auto-Submit: Robert Findley <rfindley@google.com>