telemetry: acquire upload token before starting uploader child
Instead of acquiring the upload token after the child is started,
acquire it before starting the child. Then, if the token isn't acquired
and crashmonitoring isn't requested, the child doesn't need to be
started at all. The GO_TELEMETRY_UPLOAD environment variable is set to
indicate to the child that uploading was requested. We need to set the
environment variable (or add some other way to communicate between the
parent and child such as a command line argument) because if
crashmonitoring was also requested the child would have no way of
knowing if the upload token was acquired.

The X_TELEMETRY_CHILD environment variable is renamed to
GO_TELEMETRY_CHILD to make it clear the variable comes from the go
toolchain.

Fixes golang/go#67700

Change-Id: Ifc0b4ecd8e77711bef1fcb423c96a574aec3ae09
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/589375
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>