telemetry: add MaybeChild, which executes child logic if applicable
In the Go command, we can't call telemetry.Start as the first thing when
the Go command starts because we can't do significant work before
doing toolchain switching. On the other hand, we do want to run the
child logic if the Go command is running as the telemetry child as the
first thing before doing toolchain selection: the parent has already
been started using the switched toolchain.

Add a telemetry.MaybeChild function that executes the child code if the
go command is running as the child that will be called immediately when
the go command starts. Start will continue to check if it's the child
for now to avoid breaking current usages of it but we might want to
remove that code from Start in the future to make things a bit simpler.

The initialization of the parent/child process that was done by start
has now been moved into parent and child. There is limited duplication
(for example, initializing the telemetry directory) but much of the work
is different between them. For instance, we don't need to check the mode
in the child function because it was already checked by the parent
before starting the child.

One unfortunate consequence of doing this is that the manipulation and
checking of the GO_TELEMETRY_CHILD environment variable now happens in
four places (EntryPoint, Start, parent, and startChild) rather than two
previously (Start and parent)

Change-Id: I3d31e44ee3a141b08ffd87f037f1a03f0513c1cc
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/592017
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>