counter: add OpenDir to open counters in the given telemetry directory
This is needed so that the go toolchain commands can only depend on the
counter package and not the top-level telemetry package. The top-level
telemetry package pulls in extra code (for uploading for example) that
is increasing toolchain binary sizes.

For #68109

Change-Id: I87e6526e17d726ca7a7ba35a6802bc914f87ca65
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/593995
Auto-Submit: Michael Matloob <matloob@golang.org>
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>