internal/telemetry: avoid use of x/mod/module.IsPseudoVersion
x/mod/module.IsPseudoVersion pulls in regexp, which pulls in unicode.
CL 547878 introduced this usage to enable tracking of prerelease
versions. But the goal could've achieved with just simply tweaking the
previous heuristic. Adjust the heuristic to count the number of "-"
and add the comment.

For golang/go#68109

Change-Id: Icd902dc460b489fd1eb25839cc08d90fd8bac8f3
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/594335
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>