cmd/gotelemetry: fix chart x-axis time interval display
Align the x-axis interval to match the report interval,
by setting the observablehq plot library's interval option
to be the upload date's weekday.

https://observablehq.com/plot/features/intervals#timeInterval

And use the time format string that specifies the time
zone, to avoid misinterpretation in Javascript.

The Week field of telemetry.Report is meant to be
the start of the week period. But, uploader and
the gotelemetry view's active counter aggregation
had been using the field to store the end of
the week period.

Update the comment of telemetry.Report to match
how we use the field.

In chart drawing, it is easier to work with the
start day of each week. The report with week `w`
to the half-open `[w-7days, w)` period.
Presenting the data with the bin that corresponds
to [w-7days, w) is more compelling than mapping
it to [w, w+7days). So, use w-7days as the Week
field of the datum (the data entries for charts).

Fixes golang/go#68206

Change-Id: I0db06f44a06bd7017cf0e0febd3f555bd1e201d5
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/595962
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Hyang-Ah Hana Kim <hyangah@gmail.com>
Reviewed-by: Robert Findley <rfindley@google.com>