internal/upload: use UTC time in TestRun_Concurrent
When (1) UTC time is in the next day, and (2) weekends day is exactly
right, the first count in TestRun_Concurrent will fall to be aggregated
into a report whose end date is too old. As a result, we can expect
approximately a (7/24)*(1/7) = 4% flake rate.

Fix this by correcting the counter time to UTC and pinning the upload
start time, so that the first count will not be on an earlier day, and
we avoid date skew.

Fixes golang/go#67737

Change-Id: If3b99afe4cbb6d12aa3e87177fc663f4c20ebd12
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/596316
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>