internal/telemetry: use devel for go versions with experiments
If Go is built with a GOEXPERIMENT, the version it reports will contain
an "X:" compenent for the experiment. So for instance, if go1.23rc2
(with a VERSION file specifying go1.23rc2) is built with
GOEXPERIMENT=aliastypeparams, it will report go version
"go1.23rc2 X:aliastypeparams". It should report devel instead because
the build differs from the standard 1.23rc2 build. Further, the go
version is embedded in the counter file names and the "X:" component
causes issues on Windows.

So instead if there's an "X:" component in the go version,
internal/telemetry.ProgInfo will return goVers devel.

For #68579

Change-Id: Iac405a7e2eda0d0ee4c66ffb8591b6874e7c1723
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/603855
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>