internal/upload: only download the upload config if necessary
Refactor the upload pass so that the upload config is downloaded lazily,
and only if an upload is actually going to be performed. More
refactoring is desirable, but this change is intentionally minimal as it
will be cherry-picked into Go 1.23.1 to address golang/go#68946. TODOs
are left for future CLs.

For golang/go#68946

Change-Id: I4c565b333d92361a8380dd790b1a4f2d7223a37c
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/607796
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Robert Findley <rfindley@google.com>