godev/cmd/worker: don't include buckets in the chart ID
Right now, when you click on a specific chart in the chart browser, it
updates the location to include the full counter line, including
buckets. The buckets are suppressed in the chart name only due to a
template func that trims them.

But buckets are not relevant to the identity of the chart, and result in
a very long URL. Trim the ID to only include the counter prefix.

Change-Id: Iddd1a932df4f44beb490e7225208d9c476ca7e56
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/611841
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Hongxiang Jiang <hxjiang@golang.org>