cmd/stacks: move from x/tools@effd83eb
Since the stacks command now serves not only gopls but the
go command, the compiler, and delve, it belongs in this repo.

The tests were broken at tip due to the recent pre-go1.25
change in sizeof(token.FileSet), which caused older gopls
not to build because of the missing GoVersion info.
This CL includes a fix and an assertion.

Change-Id: I9ed7894cfb6111c0843d78d658add29e904e8bf4
Reviewed-on: https://go-review.googlesource.com/c/telemetry/+/678275
Auto-Submit: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>